WIP moving e2e macro code to e2e macro crate
