+++ crates/e2e/Cargo.toml
@@ -15,6 +15,7 @@ categories = ["no-std", "embedded"]
 include = ["/Cargo.toml", "src/**/*.rs", "/README.md", "/LICENSE"]
 
 [dependencies]
+ink_e2e_macro = { version = "4.0.0-alpha.3", path = "./macro" }
 ink_env = { version = "4.0.0-alpha.3", path = "../env" }
 
 contract-metadata = { version = "2.0.0-alpha.2" }
