+++ crates/e2e/macro/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "ink_e2e_macro"
+version = "4.0.0-alpha.3"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2021"
+
+license = "Apache-2.0"
+readme = "README.md"
+repository = "https://github.com/paritytech/ink"
+documentation = "https://docs.rs/ink_macro/"
+homepage = "https://www.parity.io/"
+description = "[ink!] Macro for generating end-to-end tests"
+keywords = ["wasm", "parity", "webassembly", "blockchain", "edsl"]
+categories = ["no-std", "embedded"]
+include = ["Cargo.toml", "src/**/*.rs", "README.md", "LICENSE"]
+
+[lib]
+name = "ink_e2e_macro"
+proc-macro = true
+
+[dependencies]
+derive_more = "0.99.17"
+env_logger = "0.9.1"
+log = "0.4.17"
+serde_json = "1.0.85"
+syn = "1"
+proc-macro2 = "1"
+quote = "1"
