+++ crates/ink/codegen/src/lib.rs
@@ -60,10 +60,6 @@ impl<'a> CodeGenerator for &'a ir::InkTraitDefinition {
     type Generator = generator::TraitDefinition<'a>;
 }
 
-impl<'a> CodeGenerator for &'a ir::InkE2ETest {
-    type Generator = generator::InkE2ETest<'a>;
-}
-
 impl<'a> CodeGenerator for &'a ir::InkTest {
     type Generator = generator::InkTest<'a>;
 }
