+++ crates/ink/ir/src/ir/mod.rs
@@ -19,9 +19,7 @@ mod blake2;
 mod chain_extension;
 mod config;
 mod contract;
-mod e2e_config;
 mod idents_lint;
-mod ink_e2e_test;
 mod ink_test;
 mod item;
 mod item_impl;
@@ -69,8 +67,6 @@ pub use self::{
     },
     config::Config,
     contract::Contract,
-    e2e_config::E2EConfig,
-    ink_e2e_test::InkE2ETest,
     ink_test::InkTest,
     item::{
         Event,
