+++ crates/ink/ir/src/lib.rs
@@ -51,7 +51,6 @@ pub use self::{
         Event,
         ExtensionId,
         ImplItem,
-        InkE2ETest,
         InkItem,
         InkItemTrait,
         InkTest,
